ai_model_project/requirements.txt
flask
tensorflow
joblib
pillow
